style: Add private/public to class properties and methods, 1 class per file
